requirement.txt
numpy
tensorflow
matplotlib
fastapi
uvicorn
python-multipart
pillow